refactor: 출석 체크 api 분리
